Adapt SEO component.
- Add react-helmet.
