+++ package.json
@@ -39,6 +39,7 @@
     "react": "^16.8.0",
     "react-dom": "^16.8.0",
     "react-ga": "^2.5.7",
+    "react-helmet": "^5.2.1",
     "react-hot-loader": "^4.8.0",
     "react-jazzicon": "^0.1.3",
     "react-localization": "^1.0.13",
